% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_b_call}
\alias{build_b_call}
\title{Create an API call for B-data flows}
\usage{
build_b_call(data_item, api_key, settlement_date = NULL,
  settlement_period = NULL, year = NULL, month = NULL, week = NULL,
  process_type = NULL, start_time = NULL, end_time = NULL,
  start_date = NULL, end_date = NULL, service_type = "csv",
  version = "v1")
}
\arguments{
\item{data_item}{character string; the id of the B flow}

\item{api_key}{character string; api key retreived from the Elexon portal}

\item{...}{other values to be passed to the function}
}
\value{
url; created url for the call
}
\description{
Create an API call for B-data flows
}
