% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_remit_call}
\alias{build_remit_call}
\title{Create an API call for REMIT flows}
\usage{
build_remit_call(data_item, api_key, event_start = NULL,
  event_end = NULL, publication_from = NULL, publication_to = NULL,
  participant_id = NULL, asset_id = NULL, event_type = NULL,
  fuel_type = NULL, message_type = NULL, message_id = NULL,
  unavailability_type = NULL, active_flag = NULL, sequence_id = NULL,
  service_type = "xml", api_version = "v1")
}
\arguments{
\item{data_item}{character string; the id of the REMIT flow}

\item{api_key}{character string; api key retreived from the Elexon portal}

\item{event_start}{character string; event start (automatically cleaned by format_date)}

\item{event_end}{character string; event end (automatically cleaned by format_date)}

\item{publication_from}{character string; publication from (automatically cleaned by format_date)}

\item{publication_to}{character string; publication to (automatically cleaned by format_date)}

\item{participant_id}{character string; participant id}

\item{asset_id}{character string; asset id}

\item{event_type}{character string; event type}

\item{fuel_type}{character string; fuel type}

\item{message_type}{character string; message type}

\item{message_id}{character string; message id}

\item{unavailability_type}{character string; unavailability type}

\item{active_flag}{character string; active flag}

\item{sequence_id}{character string; sequence id}

\item{service_type}{character string; file format (csv or xml)}

\item{api_version}{character string; version of the api to use (currently on v1)}
}
\value{
string; created url for the call
}
\description{
Create an API call for REMIT flows
}
\examples{
build_remit_call(data_item = "MessageListRetrieval", api_key = "12345", settlement_date = "14-12-2016")
}
