% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_data_item}
\alias{check_data_item}
\title{Check the data item to ensure that it is a valid request}
\usage{
check_data_item(
  data_item,
  type = c("any", "B Flow", "Legacy", "REMIT"),
  silent = FALSE
)
}
\arguments{
\item{data_item}{character; the data item to check}

\item{type}{character; the type of data_item - one of "B Flow", "Legacy", or "REMIT" or "any" for any type}

\item{silent}{boolean; whether to show a warning if not a valid data item}
}
\value{
boolean: returns true if data_item is valid, false if it is not
}
\description{
Check the data item to ensure that it is a valid request
}
\examples{
check_data_item("B1720", "B Flow") #valid
check_data_item("B1720", "Legacy") #invalid - incorrect type
check_data_item("B1111", "REMIT") #invalid - incorrect data item and type
}
