% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_export_util.R
\name{check_period}
\alias{check_period}
\title{Check the the provided Settlement Period value is valid}
\usage{
check_period(period)
}
\arguments{
\item{period}{numeric/character; value to check. Must be numeric and between 1 and 50 or a character that's "*"}
}
\value{
character; period as character
}
\description{
Currently accepted values for Settlement Period is 1-50 and *
}
