% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{fix_all_parameters}
\alias{fix_all_parameters}
\title{Fixes multiple parameters}
\usage{
fix_all_parameters(params = list())
}
\arguments{
\item{params}{list; list of the params. Should have a name and a value}
}
\value{
list; list of the fixed parameters
}
\description{
Provided with a list of \code{build_..._call()} parameters, this function will fix each one and return a new list with the fixed parameters.
This is implemented by applying the \code{fix_parameter} function
}
\seealso{
fix_parameter
}
