% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generation_dataset_example}
\alias{generation_dataset_example}
\title{An example dataset from BMRS showing generation by fuel type.}
\format{A data frame with 8655 rows and 6 variables:
\describe{
  \item{settlement_date}{Settlement Date of the observation}
  \item{settlement_period}{Settlement Period of the observation}
  ...
}}
\source{
\url{ttps://www.bmreports.com/bmrs/?q=help/about-us}
}
\usage{
generation_dataset_example
}
\description{
A dataset containing UK generation by fuel type between 1 July 2019 and 3 July 2019 at half-hourly intervals.
}
\keyword{datasets}
