% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_b}
\alias{get_b}
\title{Create B flow API call and retrieve the results}
\usage{
get_b(data_item, api_key, settlement_date = NULL,
  settlement_period = NULL, year = NULL, month = NULL, week = NULL,
  process_type = NULL, start_time = NULL, end_time = NULL,
  start_date = NULL, end_date = NULL, service_type = "csv",
  version = "v1")
}
\arguments{
\item{data_item}{character string; the id of the B flow}

\item{api_key}{character string; api key retreived from the Elexon portal}
}
\value{
xml/tibble; results returned either via xml or csv depending on service_type
}
\description{
Create B flow API call and retrieve the results
}
