% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_cleaning_function}
\alias{get_cleaning_function}
\title{Get the cleaning function required for a parameter}
\usage{
get_cleaning_function(parameter, format = c("argument", "url"))
}
\arguments{
\item{parameter}{character; name of the parameter. Either the parameter as it's passed to the \code{build()} functions or the name of the parameter in the URL
depending on the value of \code{format}}

\item{format}{character; what format is \code{parameter} in? One of "argument" (default) or "url"}
}
\value{
character; name of the cleaning function
}
\description{
Before a parameter can be added to a request, it often needs to be cleaned. This function returns the appropriate function for a parameter.
Parameters can be supplied with their name used in the \code{build()} functions ("argument") or in the URL
}
