% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_column_names}
\alias{get_column_names}
\title{Get the column names for a returned csv dataset}
\usage{
get_column_names(data_item)
}
\arguments{
\item{data_item}{character string; data item for the dataset}
}
\value{
vector; a vector of character strings with the column headings
}
\description{
Get the column names for a returned csv dataset
}
\examples{
get_column_names("TEMP")
}
