% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_data_item_type}
\alias{get_data_item_type}
\title{Get the data item type of a data item}
\usage{
get_data_item_type(data_item)
}
\arguments{
\item{data_item}{character string; data item to be retrieved}
}
\description{
Get the data item type of a data item
}
\examples{
get_function("TEMP")
}
