% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_function}
\alias{get_function}
\title{Get the correct function to create the API call depending on the data item}
\usage{
get_function(data_item)
}
\arguments{
\item{data_item}{character string; data item to be retrieved}
}
\value{
function
}
\description{
Get the correct function to create the API call depending on the data item
}
\examples{
get_function("TEMP")
}
