% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_parameters}
\alias{get_parameters}
\title{TO DO: Get the required parameters for a data item}
\usage{
get_parameters(data_item)
}
\arguments{
\item{data_item}{character; the data item to get the parameters for}
}
\value{
A list containing the named parameters required for that call
}
\description{
TO DO: Get the required parameters for a data item
}
