% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_clean_csv}
\alias{parse_clean_csv}
\title{Parse a 'clean' .csv response}
\usage{
parse_clean_csv(content)
}
\arguments{
\item{content}{character; the original response object parsed as a single text string.}
}
\value{
tibble; a tibble containing the data in the .csv file
}
\description{
Some .csv files are returned without the EOF tag and with only 1 line before the data. This
function is used to parse these files, whereas the \code{parse_eof_csv()} function is used
to parse those files with the EOF tag and junk lines.
}
\seealso{
Other parsers: 
\code{\link{parse_eof_csv}()}
}
\concept{parsers}
