% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_eof_csv}
\alias{parse_eof_csv}
\title{Parse a .csv response with a EOF tag left in}
\usage{
parse_eof_csv(content)
}
\arguments{
\item{content}{character; the original response object parsed as a single text string.}
}
\value{
tibble; a tibble containing the data in the .csv file
}
\description{
Some .csv files returned from the API still have an EOF tag left at the bottom and contain 4 lines of nonsense.
This function is used to parse these files, whereas the \code{parse_clean_csv()} function is used to
parse .csv files without this tag and the junk lines.
}
\seealso{
Other parsers: 
\code{\link{parse_clean_csv}()}
}
\concept{parsers}
