% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_receive.R
\name{send_request}
\alias{send_request}
\title{Send an API request (basically a wrapper to httr:GET that adds a marker for the data item)}
\usage{
send_request(url, data_item, config_options = list())
}
\arguments{
\item{url}{A string with the API request (usually generated from build_b_call())}
}
\value{
A response() object
}
\description{
Send an API request (basically a wrapper to httr:GET that adds a marker for the data item)
}
