% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{try_parse}
\alias{try_parse}
\title{Wrapper to the tryCatch version to be used for the parsing function}
\usage{
try_parse(expr, error_message, ...)
}
\arguments{
\item{expr}{expression; expression to be evaluated for errors}

\item{error_message}{character; character string to be displayed as a warning on error}

\item{...}{extra parameters to be passed to the \code{tryCatch()} function.}
}
\value{
evaluated expression on success or empty tibble on error
}
\description{
This simple wrapper returns an empty tibble on error and returns a custom warning message.
}
