% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_call}
\alias{build_call}
\title{Build an API call (uses the appropriate function based on the data item)}
\usage{
build_call(...)
}
\description{
Build an API call (uses the appropriate function based on the data item)
}
