% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{clean_date_columns}
\alias{clean_date_columns}
\title{Reformat date, time, and datetime columns}
\usage{
clean_date_columns(x)
}
\arguments{
\item{x}{tibble/df; dataset with the columns to be formatted}
}
\value{
tibble/df; dataset with reformatted columns (if any needed reformatting)
}
\description{
Reformat date, time, and datetime columns
}
\examples{
generation_dataset_unclean <- as.data.frame(apply(generation_dataset_example, 2, as.character)) #Create a version of the example generation dataset with character columns
clean_date_columns(generation_dataset_unclean)
}
