% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{full_request}
\alias{full_request}
\title{Create an API call, send the request and retrieve the results, and parse them}
\usage{
full_request(...)
}
\value{
A tibble if service_type = "csv", otherwisea list
}
\description{
Create an API call, send the request and retrieve the results, and parse them
}
\examples{
full_request(data_item = "B1730", api_key = "12345", settlement_date = "14-12-2016")
}
