% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_parameters}
\alias{get_parameters}
\title{Get the required parameters for a data item}
\usage{
get_parameters(data_item)
}
\arguments{
\item{data_item}{string of the data item to get the parameters for}
}
\value{
A list containing the named parameters required for that call
}
\description{
Get the required parameters for a data item
}
