% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_response}
\alias{parse_response}
\title{Parse the results of a call}
\usage{
parse_response(response, format, clean_dates = TRUE)
}
\arguments{
\item{response}{A response object returned from the API request}

\item{format}{character string; format of the content of the response object; either "csv" or "xml"}

\item{clean_dates}{boolean; whether to clean date/time columns}
}
\value{
A tibble if format == "csv", otherwise a list
}
\description{
Parse the results of a call
}
\examples{
list_example <- parse_response(
send_request(
build_call("TEMP", api_key = "12345", from_date = "01 Jun 2019",
to_date = "10 Jun 2019", service_type = "xml")
), "xml")
}
