% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{send_request}
\alias{send_request}
\title{Send an API request (basically a wrapper to httr:GET)}
\usage{
send_request(url)
}
\arguments{
\item{url}{A string with the API request (usually generated from build_b_call())}
}
\value{
A response() object
}
\description{
Send an API request (basically a wrapper to httr:GET)
}
