% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_receive.R
\name{send_request}
\alias{send_request}
\title{Send an API request (basically a wrapper to httr:GET that adds a marker for the data item)}
\usage{
send_request(request, config_options = list())
}
\arguments{
\item{request}{list; a named list with at least a url to be sent and the data item contained within (most easily generated from build_call())}

\item{config_options}{list; a named list of config options to be passed to httr::GET}
}
\value{
A response() object with an added data_item attribute
}
\description{
Send an API request (basically a wrapper to httr:GET that adds a marker for the data item)
}
\examples{
send_request(
build_call(data_item = "TEMP", from_date = "01 Jun 2019", to_date = "10 Jun 2019", api_key = "test")
)
}
